Optimize mylib

+++ lib/my/my_nbr_len.c
@@ -9,9 +9,9 @@
 
 #include "my.h"
 
-int my_i64_len_base(int64_t n, const char *base)
+size_t my_i64_len_base(int64_t n, const char *base)
 {
-    int i = (n <= 0);
+    size_t i = (n <= 0);
     int64_t len = my_strlen(base);
 
     for (; n; n /= len)
@@ -19,14 +19,14 @@ int my_i64_len_base(int64_t n, const char *base)
     return i;
 }
 
-int my_i64_len(int64_t n)
+size_t my_i64_len(int64_t n)
 {
     return my_i64_len_base(n, BASE_DEC);
 }
 
-int my_u64_len_base(uint64_t n, const char *base)
+size_t my_u64_len_base(uint64_t n, const char *base)
 {
-    int i = 0;
+    size_t i = 0;
     uint64_t len = my_strlen(base);
 
     for (; n; n /= len)
@@ -34,9 +34,9 @@ int my_u64_len_base(uint64_t n, const char *base)
     return i;
 }
 
-int my_u64_len(uint64_t n)
+size_t my_u64_len(uint64_t n)
 {
-    int i = (n == 0);
+    size_t i = (n == 0);
 
     for (; n; n /= 10) {
         i++;
@@ -44,7 +44,7 @@ int my_u64_len(uint64_t n)
     return i;
 }
 
-int my_nbr_len_base(int n, const char *base)
+size_t my_nbr_len_base(size_t n, const char *base)
 {
     return my_i64_len_base((int)n, base);
 }
